deduplicate unit test code + another round of style fixes
