+++ .travis.yml
@@ -64,16 +64,16 @@ before_install:
   if [[ "$USE_DOCKER" == "TRUE" ]]; then
   docker run -it -v $(pwd):/root/repo nnvmc/base /bin/bash -c "cd /root/repo && git clone https://github.com/DCM-UPB/MCIntegratorPlusPlus.git && cp config.sh MCIntegratorPlusPlus/ && cd MCIntegratorPlusPlus && ./build.sh" || exit 1;
   else
-  git clone https://github.com/DCM-UPB/MCIntegratorPlusPlus.git && cp config.sh MCIntegratorPlusPlus/ && cd MCIntegratorPlusPlus && ./build.sh || exit 1;
+  git clone https://github.com/DCM-UPB/MCIntegratorPlusPlus.git && cp config.sh MCIntegratorPlusPlus/ && cd MCIntegratorPlusPlus && ./build.sh && cd .. || exit 1;
   fi;
 
 - |
   if [[ "$USE_DOCKER" == "TRUE" ]]; then
   docker run -it -v $(pwd):/root/repo nnvmc/base /bin/bash -c "cd /root/repo && git clone https://github.com/DCM-UPB/NoisyFunMin.git && cp config.sh NoisyFunMin/ && cd NoisyFunMin && ./build.sh" || exit 1;
   else
-  git clone https://github.com/DCM-UPB/NoisyFunMin.git && cp config.sh NoisyFunMin/ && cd NoisyFunMin && ./build.sh || exit 1;
+  git clone https://github.com/DCM-UPB/NoisyFunMin.git && cp config.sh NoisyFunMin/ && cd NoisyFunMin && ./build.sh && cd .. || exit 1;
   fi;
-  
+
 script:
 - |
   if [[ "$USE_DOCKER" == "TRUE" ]]; then
