+++ examples/ex7/main.cpp
@@ -43,7 +43,7 @@ protected:
     double _b;
 
 public:
-    Gaussian1D1POrbital(const double b):
+    explicit Gaussian1D1POrbital(const double b):
     WaveFunction(1, 1, 1, 1, false, false, false){
         _b=b;
     }
