+++ include/vmc/StochasticReconfigurationMCObservable.hpp
@@ -52,7 +52,6 @@ public:
             }
         }
     }
-
 };
