+++ test/main.cpp
@@ -5,121 +5,7 @@
 #include "vmc/Hamiltonian.hpp"
 #include "vmc/VMC.hpp"
 
-
-
-
-class QuadrExponential1D1POrbital: public WaveFunction
-{
-protected:
-    double _a, _b;
-
-public:
-    QuadrExponential1D1POrbital(const double a, const double b): WaveFunction(1, 1, 1, 2, true, false, false)
-        {_a=a; _b=b;}
-
-    void setVP(const double *in)
-    {
-        _a=in[0];
-        //if (_a<0.01) _a=0.01;
-        _b=in[1];
-        //if (_b<0.01) _b=0.01;
-        using namespace std;
-        //cout << "change a and b! " << _a << "   " << _b << endl;
-    }
-    void getVP(double *out)
-    {
-        out[0]=_a;
-        out[1]=_b;
-    }
-
-    void samplingFunction(const double *in, double *out)
-    {
-        *out = -2.*(_b*(in[0]-_a)*(in[0]-_a));
-        std::cout << "out " << *out << std::endl;
-    }
-
-    double getAcceptance(const double * protoold, const double * protonew)
-    {
-        return exp(protonew[0]-protoold[0]);
-    }
-
-    void computeAllDerivatives(const double *in){
-        _setD1DivByWF(0, -2.*_b*(in[0]-_a));
-        _setD2DivByWF(0, -2.*_b + (-2.*_b*(in[0]-_a))*(-2.*_b*(in[0]-_a)));
-        if (hasVD1()){
-            _setVD1DivByWF(0, 2.*_b*(in[0]-_a));
-            _setVD1DivByWF(1, -(in[0]-_a)*(in[0]-_a));
-        }
-    }
-
-    double computeWFValue(const double * protovalues)
-    {
-        return exp(0.5*protovalues[0]);
-    }
-};
-
-
-
-class Gaussian1D1POrbital: public WaveFunction
-{
-protected:
-    double _b;
-
-public:
-    explicit Gaussian1D1POrbital(const double b):
-    WaveFunction(1, 1, 1, 1, false, false, false){
-        _b=b;
-    }
-
-    void setVP(const double *in)
-    {
-        _b=*in;
-        //if (_b<0.01) _b=0.01;
-        using namespace std;
-        //cout << "change b! " << _b << endl;
-    }
-    void getVP(double *out)
-    {
-        *out=_b;
-    }
-
-    void samplingFunction(const double *in, double *out)
-    {
-        *out=-2.*_b*(*in)*(*in);
-    }
-
-    double getAcceptance(const double * protoold, const double * protonew)
-    {
-        return exp(protonew[0]-protoold[0]);
-    }
-
-    void computeAllDerivatives(const double *in){
-        _setD1DivByWF(0, -2.*_b*(*in));
-        _setD2DivByWF(0, -2.*_b+4.*_b*_b*(*in)*(*in));
-        if (hasVD1()){
-            _setVD1DivByWF(0, (-(*in)*(*in)));
-        }
-    }
-
-    double computeWFValue(const double * protovalues)
-    {
-        return exp(0.5*protovalues[0]);
-    }
-};
-
-
-
-class HarmonicOscillator1D1P: public Hamiltonian
-{
-protected:
-    double _w;
-public:
-    HarmonicOscillator1D1P(const double w, WaveFunction * wf): Hamiltonian(1, 1, wf) {_w=w;}
-    double localPotentialEnergy(const double *r)
-    {
-        return (0.5*_w*_w*(*r)*(*r));
-    }
-};
+#include "TestVMCFunctions.hpp"
 
 
 int main(){
