+++ test/ut2/main.cpp
@@ -13,10 +13,10 @@ int main(){
     const double TINY = 0.01;
     const int NTEST = 10;
 
-    EuclideanMetric * em = new EuclideanMetric(NSPACEDIM);
+    EuclideanMetric em(NSPACEDIM);
 
-    double * x = new double[NSPACEDIM];
-    double * y = new double[NSPACEDIM];
+    double x[NSPACEDIM];
+    double y[NSPACEDIM];
 
     for (int i=0; i<NSPACEDIM; ++i){
         x[i] = 2.;
@@ -26,15 +26,15 @@ int main(){
 
 
     // --- check that the distance is computed correctly
-    assert( em->dist(x, y) == 3.*sqrt(NSPACEDIM) );
-    assert( em->dist(y, x) == 3.*sqrt(NSPACEDIM) );
+    assert( em.dist(x, y) == 3.*sqrt(NSPACEDIM) );
+    assert( em.dist(y, x) == 3.*sqrt(NSPACEDIM) );
 
 
 
     // --- check derivatives
     // array that will store the analytical derivatives
-    double * analderivxy = new double[2*NSPACEDIM];
-    double * analderivyx = new double[2*NSPACEDIM];
+    double analderivxy[2*NSPACEDIM];
+    double analderivyx[2*NSPACEDIM];
 
     // random generator
     random_device rdev;
@@ -54,20 +54,20 @@ int main(){
         }
 
         // compute the distance xy
-        const double f = em->dist(x, y);
+        const double f = em.dist(x, y);
 
 
 
         // --- check the first derivative
 
-        em->distD1(x, y, analderivxy);
-        em->distD1(y, x, analderivyx);
+        em.distD1(x, y, analderivxy);
+        em.distD1(y, x, analderivyx);
 
         // check derivative in respect to x
         for (int i=0; i<NSPACEDIM; ++i){
             const double origx = x[i];
             x[i] += DX;
-            const double fdx = em->dist(x, y);
+            const double fdx = em.dist(x, y);
             const double numderiv = (fdx - f)/DX;
 
             // cout << "analderivxy[" <<  i << "] = " << analderivxy[i] << endl;
@@ -82,7 +82,7 @@ int main(){
         for (int i=0; i<NSPACEDIM; ++i){
             const double origy = y[i];
             y[i] += DX;
-            const double fdy = em->dist(x, y);
+            const double fdy = em.dist(x, y);
             const double numderiv = (fdy - f)/DX;
 
             // cout << "analderivyx[" <<  i << "] = " << analderivyx[i] << endl;
@@ -97,16 +97,16 @@ int main(){
 
         // --- check the second derivative
 
-        em->distD2(x, y, analderivxy);
-        em->distD2(y, x, analderivyx);
+        em.distD2(x, y, analderivxy);
+        em.distD2(y, x, analderivyx);
 
         // check derivative in respect to x
         for (int i=0; i<NSPACEDIM; ++i){
             const double origx = x[i];
             x[i] += DX;
-            const double fdx = em->dist(x, y);
+            const double fdx = em.dist(x, y);
             x[i] -= 2.*DX;
-            const double fmdx = em->dist(x, y);
+            const double fmdx = em.dist(x, y);
             const double numderiv = (fdx - 2.*f + fmdx)/(DX*DX);
 
             // cout << "analderivxy[" <<  i << "] = " << analderivxy[i] << endl;
@@ -121,9 +121,9 @@ int main(){
         for (int i=0; i<NSPACEDIM; ++i){
             const double origy = y[i];
             y[i] += DX;
-            const double fdy = em->dist(x, y);
+            const double fdy = em.dist(x, y);
             y[i] -= 2.*DX;
-            const double fmdy = em->dist(x, y);
+            const double fmdy = em.dist(x, y);
             const double numderiv = (fdy - 2.*f + fmdy)/(DX*DX);
 
             // cout << "analderivyx[" <<  i << "] = " << analderivyx[i] << endl;
@@ -134,17 +134,6 @@ int main(){
             y[i] = origy;
         }
     }
-
-
-
-
-
-    // --- free resources
-    delete[] analderivxy;
-    delete[] analderivyx;
-    delete[] y;
-    delete[] x;
-    delete em;
 
     return 0;
 }
