+++ test/ut5/main.cpp
@@ -8,151 +8,7 @@
 #include <iostream>
 #include <random>
 
-
-
-class PolynomialU2: public TwoBodyPseudoPotential{
-/*
-    u(r) = a * r^2 + b * r^3
-*/
-
-private:
-    double _a, _b;
-
-public:
-    PolynomialU2(EuclideanMetric * em, double a, double b):
-    TwoBodyPseudoPotential(em, 2, true, true, true){
-        _a = a;
-        _b = b;
-    }
-    ~PolynomialU2(){}
-
-    void setVP(const double *vp){
-        _a=vp[0]; _b=vp[1];
-    }
-    void getVP(double *vp){
-        vp[0]=_a; vp[1]=_b;
-    }
-
-    double ur(const double &r){
-        return _a*r*r + _b*r*r*r;
-    }
-
-    double urD1(const double &r){
-        return 2.*_a*r + 3.*_b*r*r;
-    }
-
-    double urD2(const double &r){
-        return 2.*_a + 6.*_b*r;
-    }
-
-    void urVD1(const double &r, double * vd1){
-        vd1[0] = r*r;
-        vd1[1] = r*r*r;
-    }
-
-    void urD1VD1(const double &r, double * d1vd1){
-        d1vd1[0] = 2.*r;
-        d1vd1[1] = 3.*r*r;
-    }
-
-    void urD2VD1(const double &r, double * d2vd1){
-        d2vd1[0] = 2.;
-        d2vd1[1] = 6.*r;
-    }
-};
-
-
-
-class FlatU2: public TwoBodyPseudoPotential{
-/*
-    u(r) = K
-*/
-private:
-    double _K;
-
-public:
-    FlatU2(EuclideanMetric * em, const double &K):
-    TwoBodyPseudoPotential(em, 1, true, true, true){
-        _K = K;
-    }
-    ~FlatU2(){}
-
-    void setVP(const double *vp){
-        _K = vp[0];
-    }
-    void getVP(double *vp){
-        vp[0] = _K;
-    }
-
-    double ur(const double &r){
-        return _K;
-    }
-
-    double urD1(const double &r){
-        return 0.;
-    }
-
-    double urD2(const double &r){
-        return 0.;
-    }
-
-    void urVD1(const double &r, double * vd1){
-        vd1[0] = 1.;
-    }
-
-    void urD1VD1(const double &r, double * d1vd1){
-        d1vd1[0] = 0.;
-    }
-
-    void urD2VD1(const double &r, double * d2vd1){
-        d2vd1[0] = 0.;
-    }
-};
-
-
-class He3u2: public TwoBodyPseudoPotential{
-/*
-    u(r) = b/r^5
-*/
-
-private:
-    double _b;
-
-public:
-    explicit He3u2(EuclideanMetric * em):
-    TwoBodyPseudoPotential(em, 1, true, true, true){
-        _b = -1.;
-    }
-
-    void setVP(const double *vp){_b=vp[0];}
-    void getVP(double *vp){vp[0]=_b;}
-
-    double ur(const double &dist){
-        return _b/pow(dist, 5);
-    }
-
-    double urD1(const double &dist){
-        return -5.*_b/pow(dist, 6);
-    }
-
-    double urD2(const double &dist){
-        return 30.*_b/pow(dist, 7);
-    }
-
-    void urVD1(const double &dist, double * vd1){
-        vd1[0] = 1./pow(dist, 5);
-    }
-
-    void urD1VD1(const double &dist, double * d1vd1){
-        d1vd1[0] = -5./pow(dist, 6);
-    }
-
-    void urD2VD1(const double &dist, double * d1vd1){
-        d1vd1[0] = 30./pow(dist, 7);
-    }
-};
-
-
+#include "TestVMCFunctions.hpp"
 
 
 int main(){
@@ -195,7 +51,7 @@ int main(){
     Psi->addWaveFunction(J_4);
 
     // particles position
-    double * x = new double[NPART*NSPACEDIM];
+    double x[NPART*NSPACEDIM];
 
     // pick x from a grid
     const double K = 0.5;
@@ -211,7 +67,7 @@ int main(){
     }
 
     // variational parameters
-    double * vp = new double[Psi->getNVP()];
+    double vp[Psi->getNVP()];
     Psi->getVP(vp);
 
 
@@ -221,7 +77,7 @@ int main(){
     assert( Psi->getNVP() == J_1->getNVP()+J_2->getNVP()+J_3->getNVP()+J_4->getNVP());
     int contivp = 0;
     for (int iJ=0; iJ<4; ++iJ){
-        double * Jvp = new double[J[iJ]->getNVP()];
+        double Jvp[J[iJ]->getNVP()];
         J[iJ]->getVP(Jvp);
         for (int ivp=0; ivp<J[iJ]->getNVP(); ++ivp){
             // cout << "vp[" << contivp+ivp << "] = " << vp[contivp+ivp] << "    Jvp[" << ivp << "] = " << Jvp[ivp] << endl;
@@ -246,16 +102,15 @@ int main(){
             Jvp[ivp] = origvp;
             Psi->setVP(vp);
         }
-        delete[] Jvp;
         contivp += J[iJ]->getNVP();
     }
 
 
 
     // --- check the sampling function
-    double * protov = new double[4];
+    double protov[4];
     Psi->samplingFunction(x, protov);
-    double ** protovJ = new double*[4]; for (int i=0; i<4; ++i) protovJ[i] = new double;
+    double * protovJ[4]; for (int i=0; i<4; ++i) protovJ[i] = new double;
     for (int iJ=0; iJ<4; ++iJ){
         J[iJ]->samplingFunction(x, protovJ[iJ]);
         // cout << "Psi protovalue = " << protov[iJ] << "    J_" << iJ+1 << " protovalue = " << protovJ[iJ][0] << endl;
@@ -272,9 +127,9 @@ int main(){
         }
     }
     // compute the new protovalues
-    double * protovnew = new double[4];
+    double protovnew[4];
     Psi->samplingFunction(x, protovnew);
-    double ** protovJnew = new double*[4]; for (int i=0; i<4; ++i) protovJnew[i] = new double;
+    double * protovJnew[4]; for (int i=0; i<4; ++i) protovJnew[i] = new double;
     J_1->samplingFunction(x, protovJnew[0]);
     J_2->samplingFunction(x, protovJnew[1]);
     J_3->samplingFunction(x, protovJnew[2]);
@@ -299,7 +154,7 @@ int main(){
 
     // initial wave function
     double f, fdx, fmdx, fdvp, fdxdvp, fmdxdvp;
-    double * samp = new double[4];
+    double samp[4];
     Psi->samplingFunction(x, samp); f = exp(samp[0]+samp[1]+samp[2]+samp[3]);
 
 
@@ -419,21 +274,12 @@ int main(){
     }
 
 
-
-
-    delete[] samp;
     for(int i=0; i<4; ++i){
         delete protovJnew[i];
     }
-    delete[] protovJnew;
-    delete[] protovnew;
     for(int i=0; i<4; ++i){
         delete protovJ[i];
     }
-    delete[] protovJ;
-    delete[] protov;
-    delete[] vp;
-    delete[] x;
     delete Psi;
     delete[] J;
     delete J_4;
